fix(integration): PHP 8.4 Closure::__debugInfo() now includes the name, file, and line
