+++ tests/integration/attributes/test_transaction_closure_clm.php84.php
@@ -11,11 +11,8 @@ class name, and lineno for closures.
 
 /*SKIPIF
 <?php
-if (version_compare(PHP_VERSION, "7.0", "<")) {
-  die("skip: CLM for PHP 5 not supported\n");
-}
-if (version_compare(PHP_VERSION, "8.4", ">=")) {
-  die("skip: newer test for PHP 8.4+\n");
+if (version_compare(PHP_VERSION, "8.4", "<")) {
+  die("skip: older test for PHP 8.3 and below\n");
 }
 */
 
@@ -88,7 +85,7 @@ newrelic.code_level_metrics.enabled=true
         "guid": "??",
         "traceId": "??",
         "transactionId": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "timestamp": "??",
         "duration": "??",
         "priority": "??",
@@ -97,9 +94,9 @@ newrelic.code_level_metrics.enabled=true
       },
       {},
       {
-        "code.lineno": 154,
+        "code.lineno": 151,
         "code.filepath": "__FILE__",
-        "code.function": "{closure}"
+        "code.function": "{closure:__FILE__:??}"
       }
     ],
     [
@@ -109,7 +106,7 @@ newrelic.code_level_metrics.enabled=true
         "guid": "??",
         "traceId": "??",
         "transactionId": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "timestamp": "??",
         "duration": "??",
         "priority": "??",
@@ -118,9 +115,9 @@ newrelic.code_level_metrics.enabled=true
       },
       {},
       {
-        "code.lineno": 162,
+        "code.lineno": 159,
         "code.filepath": "__FILE__",
-        "code.function": "{closure}"
+        "code.function": "{closure:__FILE__:??}"
       }
     ],
     [
@@ -130,7 +127,7 @@ newrelic.code_level_metrics.enabled=true
         "guid": "??",
         "traceId": "??",
         "transactionId": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "timestamp": "??",
         "duration": "??",
         "priority": "??",
@@ -139,9 +136,9 @@ newrelic.code_level_metrics.enabled=true
       },
       {},
       {
-        "code.lineno": 162,
+        "code.lineno": 159,
         "code.filepath": "__FILE__",
-        "code.function": "{closure}"
+        "code.function": "{closure:__FILE__:??}"
       }
     ]
   ]
