+++ tests/integration/attributes/test_transaction_closure_clm_off.php
@@ -13,6 +13,9 @@ The agent should *not* send code level metrics (CLM) when disabled.
 if (version_compare(PHP_VERSION, "7.0", "<")) {
   die("skip: CLM for PHP 5 not supported\n");
 }
+if (version_compare(PHP_VERSION, "8.4", ">=")) {
+  die("skip: newer test for PHP 8.4+\n");
+}
 */
 
 /*INI
