+++ tests/integration/attributes/test_transaction_closure_clm_off.php84.php
@@ -5,17 +5,13 @@
  */
 
 /*DESCRIPTION
-The agent should send code level metrics (CLM) including function name,
-class name, and lineno for closures.
+The agent should *not* send code level metrics (CLM) when disabled.
  */
 
 /*SKIPIF
 <?php
-if (version_compare(PHP_VERSION, "7.0", "<")) {
-  die("skip: CLM for PHP 5 not supported\n");
-}
-if (version_compare(PHP_VERSION, "8.4", ">=")) {
-  die("skip: newer test for PHP 8.4+\n");
+if (version_compare(PHP_VERSION, "8.4", "<")) {
+  die("skip: older test for PHP 8.3 and below\n");
 }
 */
 
@@ -23,7 +19,7 @@ if (version_compare(PHP_VERSION, "8.4", ">=")) {
 newrelic.distributed_tracing_enabled=1
 newrelic.span_events_enabled=1
 newrelic.cross_application_tracer.enabled=false
-newrelic.code_level_metrics.enabled=true
+newrelic.code_level_metrics.enabled=false
 */
 
 /*EXPECT_ANALYTICS_EVENTS
@@ -88,7 +84,7 @@ newrelic.code_level_metrics.enabled=true
         "guid": "??",
         "traceId": "??",
         "transactionId": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "timestamp": "??",
         "duration": "??",
         "priority": "??",
@@ -96,11 +92,7 @@ newrelic.code_level_metrics.enabled=true
         "parentId": "??"
       },
       {},
-      {
-        "code.lineno": 154,
-        "code.filepath": "__FILE__",
-        "code.function": "{closure}"
-      }
+      {}
     ],
     [
       {
@@ -109,7 +101,7 @@ newrelic.code_level_metrics.enabled=true
         "guid": "??",
         "traceId": "??",
         "transactionId": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "timestamp": "??",
         "duration": "??",
         "priority": "??",
@@ -117,11 +109,7 @@ newrelic.code_level_metrics.enabled=true
         "parentId": "??"
       },
       {},
-      {
-        "code.lineno": 162,
-        "code.filepath": "__FILE__",
-        "code.function": "{closure}"
-      }
+      {}
     ],
     [
       {
@@ -130,7 +118,7 @@ newrelic.code_level_metrics.enabled=true
         "guid": "??",
         "traceId": "??",
         "transactionId": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "timestamp": "??",
         "duration": "??",
         "priority": "??",
@@ -138,11 +126,7 @@ newrelic.code_level_metrics.enabled=true
         "parentId": "??"
       },
       {},
-      {
-        "code.lineno": 162,
-        "code.filepath": "__FILE__",
-        "code.function": "{closure}"
-      }
+      {}
     ]
   ]
 ]
