+++ tests/integration/jit/tracing/test_span_events_are_created_upon_uncaught_handled_exception.php
@@ -15,6 +15,9 @@ error events should be created.
 <?php
 
 require('skipif.inc');
+if (version_compare(PHP_VERSION, "8.4", ">=")) {
+  die("skip: newer test for PHP 8.4+\n");
+}
 
 
 */
