+++ tests/integration/opcache/disabled/test_span_events_are_created_upon_caught_error.php84.php
@@ -15,8 +15,8 @@ should be created.
 <?php
 
 require('skipif.inc');
-if (version_compare(PHP_VERSION, "8.4", ">=")) {
-  die("skip: newer test for PHP 8.4+\n");
+if (version_compare(PHP_VERSION, "8.4", "<")) {
+  die("skip: older test for PHP 8.3 and below\n");
 }
 
 */
@@ -29,16 +29,13 @@ newrelic.cross_application_tracer.enabled = false
 display_errors=1
 log_errors=0
 error_reporting = E_ALL
-opcache.enable=1
-opcache.enable_cli=1
+opcache.enable=0
+opcache.enable_cli=0
 opcache.file_update_protection=0
 opcache.jit_buffer_size=32M
 opcache.jit=function
 */
 
-/*PHPMODULES
-zend_extension=opcache.so
-*/
 /*EXPECT_ERROR_EVENTS
 [
   "?? agent run id",
@@ -150,7 +147,7 @@ zend_extension=opcache.so
         "transactionId": "??",
         "sampled": true,
         "priority": "??",
-        "name": "Custom\/{closure}",
+        "name": "Custom\/{closure:__FILE__:??}",
         "guid": "??",
         "timestamp": "??",
         "duration": "??",
@@ -171,6 +168,7 @@ zend_extension=opcache.so
 /*EXPECT_REGEX
 ^\s*(PHP )?Fatal error:\s*foo in .*? on line [0-9]+\s*$
 */
+require('opcache_test.inc');
 
 set_error_handler(
     function (int $errno, string $errstr) {
